implement pipelines mechanism for 2d plots
