+++ src/scipp/plt/__init__.py
@@ -1,4 +1,4 @@
 # SPDX-License-Identifier: BSD-3-Clause
 # Copyright (c) 2022 Scipp contributors (https://github.com/scipp)
 
-from .plot import plot
+from .main import main as plot
