+++ src/scipp/plt/pipeline.py
@@ -1,0 +1,8 @@
+class Pipeline(list):
+    pass
+
+    def run(self, model):
+        out = model
+        for step in self:
+            out = step.func(out, step.values())
+        return out
